src/requirements.txt
gitpython
pyyaml
jinja2
jinja2-time
